requirements.txt
databricks-sdk
datacontract-cli[databricks]
